fix: address PR review — move process code to git::process, fix Windows drain
Move the process-group kill / timed-capture logic out of git::status::worktree
into a dedicated git::process module (with its tests), keeping worktree.rs
focused.

Fix a Windows regression: on non-unix, run_git_op_capturing now falls back to
Command::output(), which drains both pipes concurrently, instead of the no-op
read_ready that let a large-output git op fill the pipe, block, and be reported
as a false timeout.

Bound read_ready to a fixed number of reads per pass, make the mutating-op
timeout a fixed const (drop the [git] op_timeout_secs config), and trim the
CHANGELOG entry.
