+++ src/git/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod commit_files;
 pub(crate) mod github;
 pub(crate) mod graph;
 pub(crate) mod graph_render;
+pub(crate) mod process;
 pub(crate) mod scanner;
 pub(crate) mod status;
