+++ src/git/status/mod.rs
@@ -9,8 +9,6 @@ mod test_support;
 #[cfg(test)]
 mod tests_basic;
 #[cfg(test)]
-mod tests_fetch_kill;
-#[cfg(test)]
 mod tests_refs;
 #[cfg(test)]
 mod tests_submodule;
@@ -18,7 +16,6 @@ mod worktree;
 
 use change_summary::{ChangeSummary, collect_change_summary};
 use worktree::{collect_worktree_info, compute_ahead_behind, fetch_remote_silent};
-pub(crate) use worktree::{kill_in_flight_git_ops, run_git_op_capturing};
 
 #[derive(Clone, Debug)]
 pub(crate) struct RepoStatus {
